Created Dictionary Class wrote tests and passed for basic load and get functionality.
load() method takes in a string as a path file and loads given file.
Words are separated by EOL and are cleaned of whitespace and non alphabetic chars

If null is give to load the default file path is used to load. This path is set within the class object.