Added feature if no match can be made, a single digit can be left as is at that point. No two consecutive digits can remain unchanged
This works by checking if a word was matched we can have one number left as is before we have to find another word match.
It also allows for the first digit in the string to be left as is.
This means there will never be two digits in a row at any point.

Refacted code, added comments , added file read in through command line.

Fixed some edge case bugs
Bugs
no input was given
File directory given does not exsist.
Removed '-' if the last digit is left as is.